requirements.txt
torch==2.6.0
transformers==4.51.3
accelerate==1.7.0
datasets==4.0.0
evaluate==0.4.3
einops==0.8.1
flash_attn==2.7.4.post1
flash-linear-attention==0.2.1
huggingface-hub==0.30.2
lightning==2.5.1.post0
matplotlib==3.10.0
nvitop==1.5.0
peft==0.13.0
pytorch-lightning==2.5.1.post0
pyarrow==20.0.0
safetensors==0.5.3
sentencepiece==0.2.0
tensorboard==2.19.0
tokenizers==0.21.1
trl==0.18.1
wandb==0.19.10
xformers==0.0.30

# Optional / utility packages — keep if you use them
jsonargparse==4.39.0
loguru==0.7.3
tqdm-multiprocess
python-dotenv==1.1.1
sentry-sdk==2.27.0
